check before kill

+++ xorg/bin/refbar.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
 barpid=$(pstree -lp | grep status_right | sed "s/.*sleep(\([0-9]\+\)).*/\1/")
-kill "$barpid"
+# kill if sleeping
+if [[ -n "$barpid" && "$barpid" =~ ^[0-9]+$ ]]; then
+    kill "$barpid"
+fi
